refactor: 모든 커스텀 훅들을 hooks 폴더로 이동 Github issue #163
